job-app-chain/requirements.txt
langgraph
langgraph-cli[inmem]
langchain
langchain-google-genai
langchain-google-community
python-jobspy
python-dotenv
playwright
gspread
google-auth-oauthlib
pytest